+++ changes.rst
@@ -28,6 +28,7 @@ Changes:
    float``, and ``set of int`` types.
 -  Automatically detect SCIP 8.1.0 and potential future versions on Windows.
 -  Fix crash when using type aliases with ``-O2`` (:bugref:`783`).
+-  Automatically detect Gurobi 11.0 and potential future versions.
 
 .. _v2.8.1: